{{cookiecutter.repo_name}}/requirements.txt
mlflow[extras]==2.8.1
psycopg2
jupyterlab
jupyterlab-git
pylint
pytest
black
ydata-profiling[notebook]==4.6.2
pandas==2.0.3
matplotlib
seaborn
ipywidgets
statsmodels
scikit-learn
omegaconf
hydra-joblib-launcher
hydra-core
cookiecutter